requirements.txt
qiskit~=1.4.3
torch~=2.7.1
numpy~=2.3.1
matplotlib~=3.10.3
PyYAML~=6.0.2
tensorcircuit~=0.12.0
jax~=0.6.2
flax~=0.10.6
optax~=0.2.5